Don't run gstreamer in a sub mainloop with the same maincontext
When generating a speaker thumbnail for a slide containing a video,
you can't generate this thumbnail using Gstreamer under a callback
triggered by Clutter AND at the same time run the mainloop in the
current context because rerunning the mainloop on the same context
implies that you are going to re-enter the clutter thread lock, which
means you're going to end up in a dead lock situation.

This patch prevents the dead lock by running the gstreamer commands on
a different context (even though still on the same thread).

The drawback of this fix is that it is going block the UI rendering
until the video thumbnail has been generated (depending on the video
that can be noticable).